Change type of id in Note and Task

+++ app/src/main/java/com/example/notes/data/repositories/notes/NoteRepository.kt
@@ -7,11 +7,11 @@ interface NoteRepository {
 
     suspend fun getAllNotes(): Flow<List<Note>>
 
-    suspend fun getNoteById(noteId: Long): Note
+    suspend fun getNoteById(noteId: Int): Note
 
     suspend fun addNote(note: Note)
 
     suspend fun updateNote(note: Note)
 
-    suspend fun deleteNotes(notes: List<Long>)
+    suspend fun deleteNotes(notes: List<Int>)
 }
