+++ app/src/main/java/com/example/notes/data/room/dao/TaskDao.kt
@@ -20,5 +20,5 @@ interface TaskDao {
     suspend fun updateTask(task: Task)
 
     @Query("DELETE FROM task WHERE id IN (:tasks)")
-    suspend fun deleteTasks(tasks: List<Long>)
+    suspend fun deleteTasks(tasks: List<Int>)
 }
