+++ app/src/main/java/com/example/notes/domain/usecases/NoteUseCases.kt
@@ -16,7 +16,7 @@ class GetAllNotesInteractor(private val repository: NoteRepository) {
 }
 
 class GetNoteByIdInteractor(private val repository: NoteRepository) {
-    suspend operator fun invoke(noteId: Long) = repository.getNoteById(noteId)
+    suspend operator fun invoke(noteId: Int) = repository.getNoteById(noteId)
 }
 
 class AddNoteInteractor(private val repository: NoteRepository) {
@@ -28,5 +28,5 @@ class UpdateNoteInteractor(private val repository: NoteRepository) {
 }
 
 class DeleteNotesInteractor(private val repository: NoteRepository) {
-    suspend operator fun invoke(notes: List<Long>) = repository.deleteNotes(notes)
+    suspend operator fun invoke(notes: List<Int>) = repository.deleteNotes(notes)
 }
